Extract Bench test in its own file
