+++ examples/readme.js
